third_party/WebKit/LayoutTests/http/tests/security/isolatedWorld/bypass-main-world-csp-javascript-urls-expected.txt

ALERT: Running test #1

ALERT: Bypass main world's CSP with javascript: URL
ALERT: iframe javascript: src running
ALERT: iframe loaded
ALERT: Running test #0


This test ensures that scripts run in isolated worlds marked with their own Content Security Policy can bypass javascript: URL restrictions.
